bot/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
aiohttp==3.9.1
openai==1.12.0


